+++ src/main/java/com/example/springboot/Repo/MenuItemRepository.java
@@ -3,8 +3,7 @@ package com.example.springboot.Repo;
 import com.example.springboot.Domain.MenuItem;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MenuItemRepository extends CrudRepository<MenuItem, Integer> {
+public interface MenuItemRepository extends CrudRepository<MenuItem, Long> {
     MenuItem findByName(String name);
     Boolean existsByName(String name);
-
 }
